Changed all buffer lengths and byte counters from uint32_t and uint64_t to size_t.

+++ include/block-list-buffer.h
@@ -9,8 +9,8 @@
 // Generic Block descriptor
 struct __block_t {
     void *next;
-    uint64_t size;
-    uint64_t bytes;
+    size_t size;
+    size_t bytes;
     char *data;         // Data pointer
     void *__block_t_impl; // This is a placeholder so every implementation can be casted to this type
 };
@@ -20,12 +20,12 @@ typedef void (*block_init)(struct __block_t *);
 typedef void (*buffer_init)(IO_DESC *);
 
 // Empty List Allocation
-struct __block_t *block_list_alloc(POOL *p, uint64_t block_count);
-struct __block_t *block_list_alloc_custom(POOL *p, size_t size_of, uint64_t block_count);
+struct __block_t *block_list_alloc(POOL *p, size_t block_count);
+struct __block_t *block_list_alloc_custom(POOL *p, size_t size_of, size_t block_count);
 
 // List Data Allocation
-int block_data_alloc(POOL *p, void *block, uint64_t bytes_per_block);
-int block_data_fastalloc(POOL *p, void *block, uint64_t bytes_per_block);
+int block_data_alloc(POOL *p, void *block, size_t bytes_per_block);
+int block_data_fastalloc(POOL *p, void *block, size_t bytes_per_block);
 
 // Buffer Management
 void blb_lock(IO_HANDLE h);
@@ -36,6 +36,6 @@ int blb_init_struct(POOL *p, IO_DESC *b);
 // Buffer Access
 struct io_desc *blb_get_read_desc(IO_HANDLE h);
 struct io_desc *blb_get_write_desc(IO_HANDLE h);
-int blb_write(IO_HANDLE h, void *buf, uint64_t *len);
-int blb_read(IO_HANDLE h, void *buf, uint64_t *len);
+int blb_write(IO_HANDLE h, void *buf, size_t *len);
+int blb_read(IO_HANDLE h, void *buf, size_t *len);
 #endif
