+++ include/simple-buffers.h
@@ -8,43 +8,42 @@
 
 // Continuous Variable-Size Buffer
 struct rbiom_args {
-    uint64_t buf_bytes;
-    uint64_t block_bytes;
+    size_t buf_bytes;
+    size_t block_bytes;
     uint16_t align;
 };
 
 const IOM *get_rb_machine();
-uint64_t rb_get_size(IO_HANDLE h);
-uint64_t rb_get_bytes(IO_HANDLE h);
+size_t rb_get_size(IO_HANDLE h);
+size_t rb_get_bytes(IO_HANDLE h);
 void rbiom_update_defaults(struct rbiom_args *rb);
-const IOM *new_rb_machine(IO_HANDLE *h, uint64_t buffer_size, uint64_t block_size);
+const IOM *new_rb_machine(IO_HANDLE *h, size_t buffer_size, size_t block_size);
 
 // Fixed-size Block Buffer
 struct fbbiom_args {
-    uint64_t buf_bytes;
-    uint64_t block_bytes;
+    size_t buf_bytes;
+    size_t block_bytes;
     uint16_t align;
     uint32_t flags;
 };
 
 const IOM *get_fbb_machine();
-uint64_t fbb_get_size(IO_HANDLE h);
-uint64_t fbb_get_bytes(IO_HANDLE h);
-//const IOM *new_fbb_machine_fill(IO_HANDLE *h, uint64_t buffer_size, uint64_t block_size);
-const IOM *new_fbb_machine(IO_HANDLE *h, uint64_t buffer_size, uint64_t block_size);
+size_t fbb_get_size(IO_HANDLE h);
+size_t fbb_get_bytes(IO_HANDLE h);
+const IOM *new_fbb_machine(IO_HANDLE *h, size_t buffer_size, size_t block_size);
 
 // Sync Buffer
 // Continuous Variable-Size Buffer
 struct sync_iom_args {
-    uint64_t buf_bytes;
-    uint64_t block_bytes;
+    size_t buf_bytes;
+    size_t block_bytes;
     uint16_t align;
 };
 
 // Asynchronous Variable-Block Buffer
 struct avbb_args {
-    uint64_t block_bytes;
-    uint64_t block_count;
+    size_t block_bytes;
+    size_t block_count;
 };
 
 const IOM *get_avbb_machine();
@@ -52,9 +51,9 @@ void avbbiom_update_defaults(struct avbb_args *rb);
 
 // Asynchronous Fixed Packet Buffer
 struct afpbiom_args {
-    uint64_t header_bytes;
-    uint64_t payload_bytes;
-    uint64_t block_count;
+    size_t header_bytes;
+    size_t payload_bytes;
+    size_t block_count;
 };
 
 const IOM *get_afpb_machine();
@@ -64,8 +63,8 @@ struct afpb_packet {
     char *packet;
     char *header;
     char *payload;
-    uint64_t header_len;
-    uint64_t payload_len;
+    size_t header_len;
+    size_t payload_len;
 };
 
 // Alternate access methods
